Find tags in registry
Though based in the previous work and ideas, this commit address a
completely different use case. Here we have an admin who is using
version tags in their repo and, before updating, wants to know what are
the tags versions corresponding to the 'latest' tag in the registry. So,
instead of finding the hash for the latest in local (which probably
doesn't exist, as they are using version tags), it search in the
registry.